powergrid/pyproject.toml
[project]
name = "ai-engineer-take-home"
version = "0.1.0"
description = "Starter kit for the AI Engineer take-home exercise."
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.11",
    "numpy>=1.24",
    "requests>=2.31",
    "openrouter>=0.7.11",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "mypy",
    "ruff",
]
